# cyclotron-svc/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cyclotron API",
        "description": "Cyclotron's REST API for Dashboard management",
        "version": "1.0.0"
    },
    "host": "",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/dashboards": {
            "get": {
                "summary": "Get Dashboards",
                "description": "The Dashboards endpoint returns a list of Dashboards, excluding those that are deleted.  The list can be optionally filtered by a query string.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The q parameter can be a string value or comma-separated list of values, where each value is used as a partial match for the Dashboard name, or an exact match of one of the Dashboard's tags.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of Dashboards (without the Dashboard definition)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DashboardInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "post": {
                "summary": "Create Dashboard",
                "description": "Posting to the Dashboards endpoint creates a new Dashboard.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "dashboard",
                        "in": "body",
                        "description": "The Dashboard to be created.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The created Dashboard, including auto-populated fields such as date, rev, etc.",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Missing Dashboard name, or Dashboard definition."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboards/{name}": {
            "get": {
                "summary": "Get Dashboard",
                "description": "The Dashboard endpoint returns a given Dashboard by name.  This works regardless of whether or not the Dashboard is deleted.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single Dashboard",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "401": {
                        "description": "Authentication required: this dashboard has restricted permissions."
                    },
                    "403": {
                        "description": "View Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "put": {
                "summary": "Update Dashboard",
                "description": "Putting a modified Dashboard back to this endpoint overwrites the existing Dashboard with the modified version.  The revision counter is incremented, and a new revision is added to the /revisions endpoint.  If the Dashboard was deleted previously, it will be undeleted.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }, {
                        "name": "dashboard",
                        "in": "body",
                        "description": "The modified Dashboard.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated Dashboard, including auto-populated fields such as date, rev, etc.",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Missing Dashboard name, or Dashboard definition."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "403": {
                        "description": "Edit Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "delete": {
                "summary": "Delete Dashboard",
                "description": "Performs a soft delete, whereby the deleted flag is set to true and the Dashboard is automatically excluded from search results.  The Dashboard can still be requested directly by name, and undeleted.  The delete operation creates a new Revision.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The deleted Dashboard, including auto-populated fields such as date, rev, etc.",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Missing Dashboard name."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "403": {
                        "description": "Edit Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboards/{name}/tags": {
            "put": {
                "summary": "Update Dashboard Tags",
                "description": "The Dashboard Tags endpoint allows setting only the tags for a Dashboard.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }, {
                        "name": "tags",
                        "in": "body",
                        "description": "The modified Dashboard.",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated Dashboard, including auto-populated fields such as date, rev, etc.",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "400": {
                        "description": "Missing Dashboard tags."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "403": {
                        "description": "Edit Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboards/{name}/revisions": {
            "get": {
                "summary": "Get Dashboard Revisions",
                "description": "The Dashboard Revisions endpoint returns all the revisions for a Dashboard.  This works regardless of whether or not the Dashboard is deleted.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of Dashboard Revisions (each without the Dashboard definition)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DashboardInfo"
                            }
                        }
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboards/{name}/revisions/{revision}": {
            "get": {
                "summary": "Get Dashboard Revision",
                "description": "The Dashboard Revision endpoint returns a specifc revision for a Dashboard.  This works regardless of whether or not the Dashboard is deleted.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }, {
                        "name": "revision",
                        "in": "path",
                        "description": "The revision parameter specifies which Revision to return.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single Dashboard Revision",
                        "schema": {
                            "$ref": "#/definitions/Dashboard"
                        }
                    },
                    "401": {
                        "description": "Authentication required: this dashboard has restricted permissions."
                    },
                    "403": {
                        "description": "View Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboards/{name}/likes": {
            "get": {
                "summary": "Get Dashboard Likes",
                "description": "The Like Dashboard endpoint retrieves the list of users who have 'liked' this Dashboard.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The list of Users who like the Dashboard."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "post": {
                "summary": "Like Dashboard",
                "description": "The Like Dashboard endpoint adds the current user to the list of users who like the Dashboard.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated list of Users who like the Dashboard."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            },
            "delete": {
                "summary": "Unlike Dashboard",
                "description": "The Unlike Dashboard endpoint removes the current user from the list of users who like the Dashboard.",
                "tags": [
                    "Dashboards"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated list of Users who like the Dashboard."
                    },
                    "401": {
                        "description": "Authentication required: session key not provided."
                    },
                    "404": {
                        "description": "Dashboard not found."
                    },
                    "500": {
                        "description": "Server error."
                    }
                }
            }
        },
        "/dashboardnames": {
            "get": {
                "summary": "Get Dashboard Names",
                "description": "The Dashboard Names endpoint returns a list of the names of all non-deleted Dashboards.",
                "tags": [
                    "Dashboards"
                ],
                "responses": {
                    "200": {
                        "description": "An array of Dashboard names",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "summary": "Get Tags",
                "description": "The Tags endpoint returns a combined list of the unique Tags of all non-deleted Dashboards.",
                "tags": [
                    "Tags"
                ],
                "responses": {
                    "200": {
                        "description": "An array of Tags",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/searchhints": {
            "get": {
                "summary": "Get Search Hints",
                "description": "The Search Hints endpoint returns a combined list of the names and tags of all non-deleted Dashboards.  It is designed for autocomplete hints for Dashboard searching.",
                "tags": [
                    "Tags"
                ],
                "responses": {
                    "200": {
                        "description": "An array of Dashboard names",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/export/{name}/pdf": {
            "get": {
                "summary": "Synchronous Export to PDF",
                "description": "The Export PDF endpoint, when used synchronously, starts exporting a Dashboard to PDF and waits until it is completed and can be returned.",
                "tags": [
                    "Export"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to the generated PDF."
                    },
                    "401": {
                        "description": "Authentication required: this dashboard has restricted permissions."
                    },
                    "403": {
                        "description": "View Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "summary": "Asynchronous Export to PDF",
                "description": "The Export PDF endpoint, when used asynchronously, starts exporting a Dashboard to PDF and returns a key that can be used to poll for the status of the export.",
                "tags": [
                    "Export"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The Dashboard name.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Export job created successfully; returns the corresponding URL for /exportstatus",
                        "schema": {
                            "$ref": "#/definitions/ExportResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required: this dashboard has restricted permissions."
                    },
                    "403": {
                        "description": "View Permission denied for this Dashboard."
                    },
                    "404": {
                        "description": "Dashboard not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/exportstatus/{key}": {
            "get": {
                "summary": "Export Job Status",
                "description": "The Export Status endpoint provides the status and links to exported artifacts for an asynchronous export job.",
                "tags": [
                    "Export"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "The Export job key.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the status of the export job.",
                        "schema": {
                            "$ref": "#/definitions/ExportStatus"
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Get Users",
                "description": "The Users endpoint returns a list of all known Users in Cyclotron.",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "An array of Users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/users/{username}": {
            "get": {
                "summary": "Get User",
                "description": "The User endpoint returns the details of a specific User.",
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The sAMAccountName of the User to return.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of Users",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ExportResponse": {
            "properties": {
                "statusUrl": {
                    "description": "Link to the /exportstatus endpoint with the status for this specific export job.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "authenticatedSession": {
            "type": "apiKey",
            "description": "Session-based security using LDAP authentication",
            "name": "session",
            "in": "query"
        }
    }
}